Navigates to cart from product screen
Adds functionality to navigate to the cart screen from the product details screen.

Implements a new state and event in the ProductScreenBloc to handle the navigation.
Updates the product screen UI to dispatch the navigation event when the cart icon is pressed.

Also, makes `selectedIndex` mutable in `HomeScreen` to update the bottom navigation bar.